since its not tag specific, move the tests to a more generic spot
