+++ tests/Tags/TagsTest.php
@@ -35,90 +35,14 @@ class TagsTest extends TestCase
         $this->assertEquals($parser, $class->parser);
         $this->assertInstanceOf(TestDependency::class, $class->dependency);
     }
-
-    /** @test */
-    public function hooks_can_be_run()
-    {
-        $test = $this;
-
-        TestTags::hook('constructed', function ($payload, $next) use ($test) {
-            $test->assertEquals('initial', $payload);
-            $test->assertInstanceOf(TestTags::class, $this);
-            $this->setFoo('bar');
-
-            return $next($payload);
-        });
-
-        // Do it twice to ensure that they are executed in order
-        // and the tag is passed along through the closures.
-        TestTags::hook('constructed', function ($payload, $next) use ($test) {
-            $test->assertInstanceOf(TestTags::class, $this);
-            $test->assertEquals('initial', $payload);
-            $this->setFoo($this->foo.'baz');
-
-            return $next($payload);
-        });
-
-        $class = app(TestTags::class);
-
-        $this->assertEquals('barbaz', $class->foo);
-    }
-
-    /** @test */
-    public function hooks_from_one_tag_class_dont_happen_on_another()
-    {
-        $hooksRan = 0;
-
-        TestTags::hook('constructed', function ($payload, $next) use (&$hooksRan) {
-            $this->setFoo('bar');
-            $hooksRan++;
-
-            return $next($this);
-        });
-
-        AnotherTestTags::hook('constructed', function ($payload, $next) use (&$hooksRan) {
-            $this->setFoo($this->foo.'baz');
-            $hooksRan++;
-
-            return $next($this);
-        });
-
-        $class = app(AnotherTestTags::class);
-
-        $this->assertEquals(1, $hooksRan);
-        $this->assertEquals('baz', $class->foo);
-    }
 }
 
 class TestTags extends Tags
 {
     public $dependency;
-    public $foo = 'initial';
 
     public function __construct(TestDependency $dependency)
     {
         $this->dependency = $dependency;
-
-        $this->runHooks('constructed', $this->foo);
-    }
-
-    public function setFoo(string $foo)
-    {
-        $this->foo = $foo;
-    }
-}
-
-class AnotherTestTags extends Tags
-{
-    public $foo = '';
-
-    public function __construct()
-    {
-        $this->runHooks('constructed');
-    }
-
-    public function setFoo(string $foo)
-    {
-        $this->foo = $foo;
     }
 }
